refactor(Organizer): add mock files
