+++ src/entities/Organizer/__test__/mocks/track1Morning.js
@@ -1,0 +1,24 @@
+import { Lecture } from "../../../Lecture";
+
+export const track1Morning = [
+    new Lecture(
+        "Diminuindo tempo de execução de testes em aplicações Rails enterprise", 
+        60,
+        "Diminuindo tempo de execução de testes em aplicações Rails enterprise 60min"
+     ),
+    new Lecture(
+        "Reinventando a roda em ASP clássico", 
+        45,
+        "Reinventando a roda em ASP clássico 45min"
+     ),
+    new Lecture(
+        "Apresentando Lua para as massas", 
+        30,
+        "Apresentando Lua para as massas 30min"
+     ),
+    new Lecture(
+        "Erros de Ruby oriundos de versões erradas de gems", 
+        45,
+        "Erros de Ruby oriundos de versões erradas de gems 45min"
+    ),
+]
